Rename sentenceParts to Clauses
