+++ packages/yoastseo/src/languageProcessing/languages/es/values/Clause.js
@@ -1,26 +1,27 @@
-import SentencePart from "../../../../values/SentencePart.js";
+import Clause from "../../../../values/Clause.js";
 import getParticiples from "../helpers/internal/getParticiples.js";
 
 /**
- * Creates a Dutch-specific sentence part.
+ * Creates a Spanish-specific sentence part.
  *
  * @param {string} sentencePartText The text from the sentence part.
+ *
  * @param {Array} auxiliaries The list with auxiliaries.
  * @constructor
  */
-const DutchSentencePart = function( sentencePartText, auxiliaries ) {
-	SentencePart.call( this, sentencePartText, auxiliaries );
+const SpanishSentencePart = function( sentencePartText, auxiliaries ) {
+	Clause.call( this, sentencePartText, auxiliaries );
 };
 
-require( "util" ).inherits( DutchSentencePart, SentencePart );
+require( "util" ).inherits( SpanishSentencePart, Clause );
 
 /**
  * Returns the participles found in the sentence part.
  *
  * @returns {Array} The array of Participle Objects.
  */
-DutchSentencePart.prototype.getParticiples = function() {
+SpanishSentencePart.prototype.getParticiples = function() {
 	return getParticiples( this.getSentencePartText(), this.getAuxiliaries() );
 };
 
-export default DutchSentencePart;
+export default SpanishSentencePart;
