+++ packages/yoastseo/src/languageProcessing/languages/fr/values/Clause.js
@@ -1,27 +1,27 @@
-import SentencePart from "../../../../values/SentencePart.js";
+import Clause from "../../../../values/Clause.js";
 import getParticiples from "../helpers/internal/getParticiples.js";
 
 /**
- * Creates a Spanish-specific sentence part.
+ * Creates a French-specific sentence part.
  *
  * @param {string} sentencePartText The text from the sentence part.
- *
  * @param {Array} auxiliaries The list with auxiliaries.
+ *
  * @constructor
  */
-const SpanishSentencePart = function( sentencePartText, auxiliaries ) {
-	SentencePart.call( this, sentencePartText, auxiliaries );
+const FrenchSentencePart = function( sentencePartText, auxiliaries ) {
+	Clause.call( this, sentencePartText, auxiliaries );
 };
 
-require( "util" ).inherits( SpanishSentencePart, SentencePart );
+require( "util" ).inherits( FrenchSentencePart, Clause );
 
 /**
  * Returns the participles found in the sentence part.
  *
  * @returns {Array} The array of Participle Objects.
  */
-SpanishSentencePart.prototype.getParticiples = function() {
+FrenchSentencePart.prototype.getParticiples = function() {
 	return getParticiples( this.getSentencePartText(), this.getAuxiliaries() );
 };
 
-export default SpanishSentencePart;
+export default FrenchSentencePart;
