+++ packages/yoastseo/src/languageProcessing/languages/hu/values/Clause.js
@@ -1,26 +1,28 @@
-import SentencePart from "../../../../values/SentencePart.js";
+import Clause from "../../../../values/Clause.js";
 import getParticiples from "../helpers/internal/getParticiples.js";
 
 /**
- * Creates a English specific sentence part.
+ * Creates a Hungarian specific sentence part.
  *
  * @param {string} sentencePartText     The text from the sentence part.
  * @param {Array} auxiliaries           The list of auxiliaries from the sentence part.
+ *
  * @constructor
  */
-const EnglishSentencePart = function( sentencePartText, auxiliaries ) {
-	SentencePart.call( this, sentencePartText, auxiliaries );
+const HungarianSentencePart = function( sentencePartText, auxiliaries ) {
+	Clause.call( this, sentencePartText, auxiliaries );
 };
 
-require( "util" ).inherits( EnglishSentencePart, SentencePart );
+require( "util" ).inherits( HungarianSentencePart, Clause );
 
 /**
  * Returns the participle objects for the participles found in the sentence part.
+ *
  * @returns {Array} The list of participle objects.
  */
 
-EnglishSentencePart.prototype.getParticiples = function() {
+HungarianSentencePart.prototype.getParticiples = function() {
 	return getParticiples( this.getSentencePartText(), this.getAuxiliaries() );
 };
 
-export default EnglishSentencePart;
+export default HungarianSentencePart;
