+++ packages/yoastseo/src/languageProcessing/languages/pl/values/Clause.js
@@ -1,27 +1,26 @@
-import SentencePart from "../../../../values/SentencePart.js";
+import Clause from "../../../../values/Clause.js";
 import getParticiples from "../helpers/internal/getParticiples.js";
 
 /**
- * Creates a French-specific sentence part.
+ * Creates a Polish-specific sentence part.
  *
  * @param {string} sentencePartText The text from the sentence part.
  * @param {Array} auxiliaries The list with auxiliaries.
- *
  * @constructor
  */
-const FrenchSentencePart = function( sentencePartText, auxiliaries ) {
-	SentencePart.call( this, sentencePartText, auxiliaries );
+const PolishSentencePart = function( sentencePartText, auxiliaries ) {
+	Clause.call( this, sentencePartText, auxiliaries );
 };
 
-require( "util" ).inherits( FrenchSentencePart, SentencePart );
+require( "util" ).inherits( PolishSentencePart, Clause );
 
 /**
  * Returns the participles found in the sentence part.
  *
  * @returns {Array} The array of Participle Objects.
  */
-FrenchSentencePart.prototype.getParticiples = function() {
+PolishSentencePart.prototype.getParticiples = function() {
 	return getParticiples( this.getSentencePartText(), this.getAuxiliaries() );
 };
 
-export default FrenchSentencePart;
+export default PolishSentencePart;
